refactor: seperate header into cozy section
