+++ src/components/sections/header/cozy/style.module.scss
